refactored all Packet classes to be sealed records + serialization improvements
- added IVersionAwareSerializable to extend ISerializable with version specific logic like in the packets
- moved some core element serialization methods in their classes

fixes #68

+++ Components/MineSharp.Protocol/MinecraftClient.cs
@@ -361,10 +361,10 @@ public sealed class MinecraftClient : IAsyncDisposable, IDisposable
                                       SendPacket(new Packets.Serverbound.Play.ClientInformationPacket(
                                                      Settings.Locale,
                                                      Settings.ViewDistance,
-                                                     (int)Settings.ChatMode,
+                                                     Settings.ChatMode,
                                                      Settings.ColoredChat,
                                                      Settings.DisplayedSkinParts,
-                                                     (int)Settings.MainHand,
+                                                     Settings.MainHand,
                                                      Settings.EnableTextFiltering,
                                                      Settings.AllowServerListings)));
             }
@@ -376,10 +376,10 @@ public sealed class MinecraftClient : IAsyncDisposable, IDisposable
             _ = SendPacket(new ClientInformationPacket(
                            Settings.Locale,
                            Settings.ViewDistance,
-                           (int)Settings.ChatMode,
+                           Settings.ChatMode,
                            Settings.ColoredChat,
                            Settings.DisplayedSkinParts,
-                           (int)Settings.MainHand,
+                           Settings.MainHand,
                            Settings.EnableTextFiltering,
                            Settings.AllowServerListings));
         }
